init branch
# Conflicts:
#	contract/src/events.rs
#	contract/src/main.rs
