+++ contract/src/events/events_cep47.rs
@@ -1,0 +1,1 @@
+
